+++ src/url/parseParams.ts
@@ -1,22 +1,25 @@
-// parseParams( queryString )
-//
-// Convert queryString type Strings into a neat object
-// where each named value is an Array of URL-decoded Strings.
-// Defaults to parsing the current document URL if no paramString is passed.
-//
-// Example:
-//    var obj = parseParamS( "?foo=1&=bar&baz=&foo=2&" );
-//      ==>  {  'foo':['1','2'],  '':['bar'],  'baz':['']  }
-//    var obj = parseParamS( "" );
-//      ==>  { }
-
 type Params<P extends string> = Record<P, Array<string>>;
 
+/**
+ * Convert queryString type Strings into a neat object
+ * where each named value is an Array of URL-decoded Strings.
+ * Defaults to parsing the current document URL if no paramString is passed.
+ *
+ * Example:
+ *    var obj = parseParamS( "?foo=1&=bar&baz=&foo=2&" );
+ *      ==>  {  'foo':['1','2'],  '':['bar'],  'baz':['']  }
+ *    var obj = parseParamS( "" );
+ *      ==>  { }
+ *
+ */
 export default function parseParams<P extends string>(
-	...args: Array<string | undefined>
+	queryString?: string | undefined
 ): Params<P> {
 	const map: Record<string, Array<string>> = {};
-	const paramString = (args.length === 0 ? document.location.search : args[0] || '')
+	const paramString = (arguments.length === 0
+		? document.location.search
+		: queryString || ''
+	)
 		.trim()
 		.replace(/^[?&]/, '')
 		.replace(/&$/, '');