+++ src/dom/removeNode.ts
@@ -1,7 +1,8 @@
-export default function removeNode(node: null): null;
+export default function removeNode(node: undefined | null): null;
 export default function removeNode<N extends Node>(node: N): N;
 
-export default function removeNode<N extends Node>(node: N | null): N | null {
+export default function removeNode<N extends Node>(
+	node: N | undefined | null
 	node && node.parentNode && node.parentNode.removeChild(node);
 	return node;
 }