+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Upcoming...
 
 - <!-- Add new lines here. Version number will be decided later -->
+- feat: Allow calling `eventify()` with no arguents
 
 ## 3.6.0