+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## Upcoming...
 
-- <!-- Add new lines here. Version number will be decided later -->
+- ... <!-- Add new lines here. Version number will be decided later -->
 - feat: Allow calling `eventify()` with no arguents
 - docs: Improve docs and type signature of `parseParams()`