+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - ... <!-- Add new lines here. Version number will be decided later -->
 - feat: Allow calling `eventify()` with no arguents
 - docs: Improve docs and type signature of `parseParams()`
+- docs: Improve docs for `aquireId()`
 
 ## 3.6.0