+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Upcoming...
 
 - ... <!-- Add new lines here. Version number will be decided later -->
+- **BREAKING** feat: Rename module to `@hugsmidjan/qj`
 - feat: Add `htmlLang()` utility to resolve an Element's language
 - feat: Allow calling `eventify()` with no arguents
 - feat: `range()` should support small steps